requirements.txt
beautifulsoup4==4.12.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.9.24
charset-normalizer==2.1.1
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
oauthlib==3.2.2
packaging==23.1
pefile==2022.5.30
pip-check==2.8.1
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygsheets==2.0.6
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4
terminaltables==3.1.10
uritemplate==4.1.1
urllib3==1.26.12
